Name switch (#170)
* renamed mailbag to mailbagit

* addressed odd decode_header behavior

* updates the data after header encoding handling and turns off auto creating new data before testing
