+++ Dockerfile.production
@@ -23,4 +23,4 @@ RUN apt-get -y update && \
         https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb && \
     apt-get install -y /tmp/google-chrome-stable_current_amd64.deb && \
     \
-    pip install mailbag[pst]
+    pip install mailbagit[pst]
