+++ branching_workflow.md
@@ -1,9 +1,9 @@
-# Mailbag Branching Workflow
+# Mailbagit Branching Workflow
 
 ## Find an Issue
 
-* Review the [project](https://github.com/UAlbanyArchives/mailbag/projects/1)
-* Select an [issue](https://github.com/UAlbanyArchives/mailbag/issues)
+* Review the [project](https://github.com/UAlbanyArchives/mailbagit/projects/1)
+* Select an [issue](https://github.com/UAlbanyArchives/mailbagit/issues)
 * Identify the Issue number
 
 <img src="persona_images/issue_number.png" alt="Screenshot showing where the issue number is on Github" width="500px"/>
@@ -24,7 +24,7 @@ git branch -a
 
 ## Creating a new feature to address the issue
 
-1. Move the issue to "in progress" in the [project](https://github.com/UAlbanyArchives/mailbag/projects/1)
+1. Move the issue to "in progress" in the [project](https://github.com/UAlbanyArchives/mailbagit/projects/1)
 2. Create a new branch named `feature-[issue number]`
 
 ```bash
@@ -46,12 +46,12 @@ git push origin feature-[issue number]
 ```
 
 6. [Create a Pull Request](https://docs.github.com/en/github/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request) from your branch to develop
-7. Move the Issue in the [project](https://github.com/UAlbanyArchives/mailbag/projects/1) to "Ready for Review"
+7. Move the Issue in the [project](https://github.com/UAlbanyArchives/mailbagit/projects/1) to "Ready for Review"
 
 ## Reviewing 
 
-1. View the "Ready for Review" column in the [project](https://github.com/UAlbanyArchives/mailbag/projects/1) and identify the issue number
-2. View the original [issue](https://github.com/UAlbanyArchives/mailbag/issues) and the [pull request](https://github.com/UAlbanyArchives/mailbag/pulls)
+1. View the "Ready for Review" column in the [project](https://github.com/UAlbanyArchives/mailbagit/projects/1) and identify the issue number
+2. View the original [issue](https://github.com/UAlbanyArchives/mailbagit/issues) and the [pull request](https://github.com/UAlbanyArchives/mailbagit/pulls)
 3. Checkout the 
 
 ```bash
@@ -74,4 +74,4 @@ git checkout develop
 git merge feature-[issue number]
 ```
 
-7. Move Move the Issue in the [project](https://github.com/UAlbanyArchives/mailbag/projects/1) to "Reviewer Approved"
+7. Move Move the Issue in the [project](https://github.com/UAlbanyArchives/mailbagit/projects/1) to "Reviewer Approved"
