+++ data/mbox-sample1/1/Subject.txt
@@ -1,1 +1,1 @@
-=?utf-8?Q?The_GOP=E2=80=99s_horrendous_damage?=
+The GOP’s horrendous damage
