+++ docs/using_mailbagit/plugins.md
@@ -10,7 +10,7 @@ nav_order: 4
 
 You can make plugins for both the mailbagit format parsers and derivatives classes. This allows you to easily override the behavior for how mailbagit reads email exports or creates derivative formats. New input and derivative formats may also be provided to mailbagit to extend its functionality.
 
-The files that can be overridden can be found in the [Github repo](https://github.com/UAlbanyArchives/mailbag) in the [mailbag/formats](https://github.com/UAlbanyArchives/mailbag/tree/main/mailbag/formats) and [mailbag/derivatives](https://github.com/UAlbanyArchives/mailbag/tree/main/mailbag/derivatives) subdirectories.
+The files that can be overridden can be found in the [Github repo](https://github.com/UAlbanyArchives/mailbagit) in the [mailbagit/formats](https://github.com/UAlbanyArchives/mailbagit/tree/main/mailbagit/formats) and [mailbagit/derivatives](https://github.com/UAlbanyArchives/mailbagit/tree/main/mailbagit/derivatives) subdirectories.
 
 By default, mailbagit will look for formats in the following places:
 
@@ -21,18 +21,18 @@ By default, mailbagit will look for formats in the following places:
 		mkdir ~/myplugindir
 		mkdir ~/myplugindir/formats
 		touch ~/myplugindir/formats/pst.py
-		export MAILBAG_PLUGIN_DIR=$HOME/myplugindir
+		export MAILBAGIT_PLUGIN_DIR=$HOME/myplugindir
 	```
-2. `.mailbag/formats` and `.mailbag/derivatives` subdirectories in the user's home directory.
+2. `.mailbagit/formats` and `.mailbagit/derivatives` subdirectories in the user's home directory.
 	
 	Unix Example:
 	```
-		mkdir ~/.mailbag/derivatives
-		touch ~/.mailbag/derivatives/pdf.py
+		mkdir ~/.mailbagit/derivatives
+		touch ~/.mailbagit/derivatives/pdf.py
 	```
 
 	Example Windows path:
 	```
-	C:\Users\[my_username]\.mailbag\formats\imap.py
+	C:\Users\[my_username]\.mailbagit\formats\imap.py
 	```
-3. The formats and derivatives built into mailbag.
+3. The formats and derivatives built into mailbagit.
