+++ mailbagit/derivatives/pdf.py
@@ -1,18 +1,21 @@
 import os
 import subprocess
 import distutils.spawn
-from mailbag.derivative import Derivative
+from mailbagit.derivative import Derivative
 from structlog import get_logger
-import mailbag.helper as helper
+import mailbagit.helper as helper
 
-
+# only create format if pypff is successfully importable -
+# pst is not supported otherwise
 skip_registry = False
-
 try:
-    chromes = ["google-chrome", "chrome.exe", "chrome"]
-    chrome = next((c for c in chromes if distutils.spawn.find_executable(c)), None)
-    skip_registry = True if chrome is None else False
-
+    if distutils.spawn.find_executable("wkhtmltopdf.exe"):
+        wkhtmltopdf = "wkhtmltopdf.exe"
+    elif distutils.spawn.find_executable("wkhtmltopdf"):
+        wkhtmltopdf = "wkhtmltopdf"
+    else:
+        skip_registry = True
+        wkhtmltopdf = None
 except:
     skip_registry = True
 
@@ -20,11 +23,16 @@ log = get_logger()
 
 if not skip_registry:
 
-    class PDFChromeDerivative(Derivative):
-        derivative_name = "pdf-chrome"
+    class PDFDerivative(Derivative):
+        derivative_name = "pdf"
         derivative_format = "pdf"
-        derivative_agent = "chrome"
-        derivative_agent_version = "unknown"
+        derivative_agent = "wkhtmltopdf"
+        check_version = subprocess.Popen([wkhtmltopdf, "--version"], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        out, err = check_version.communicate()
+        if len(err) > 0:
+            log.error("Unable to access --version for " + derivative_agent + ": " + err.decode("utf-8"))
+            raise Exception("Unable to access --version for " + derivative_agent + ": " + err.decode("utf-8"))
+        derivative_agent_version = out.decode("utf-8").strip()
 
         def __init__(self, email_account, **kwargs):
             log.debug("Setup account")
@@ -53,7 +61,8 @@ if not skip_registry:
                 pdf_name = filename + ".pdf"
 
                 if message.HTML_Body is None and message.Text_Body is None:
-                    log.warn("No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ". No PDF derivative will be created.")
+                    desc = "No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ", no PDF derivative created"
+                    errors = helper.handle_error(errors, None, desc, "warn")
                 else:
                     log.debug("Writing HTML to " + str(html_name) + " and converting to " + str(pdf_name))
                     # Calling helper function to get formatted html
@@ -72,31 +81,22 @@ if not skip_registry:
                                 write_html.write(html_formatted)
                                 write_html.close()
                             command = [
-                                chrome,
-                                "--headless",
-                                "--run-all-compositor-stages-before-draw",
-                                "--disable-gpu",
-                                "--print-to-pdf-no-header",
-                                "--print-to-pdf=" + os.path.abspath(pdf_name),
+                                wkhtmltopdf,
+                                "--disable-javascript",
                                 os.path.abspath(html_name),
+                                os.path.abspath(pdf_name),
                             ]
-
-                            # Adds --no-sandbox arg to run as root in docker container if env variable set
-                            if os.environ.get("IN_CONTAINER", "").upper() == "TRUE":
-                                command.insert(4, "--no-sandbox")
-
                             log.debug("Running " + " ".join(command))
                             p = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-
                             stdout, stderr = p.communicate()
                             if p.returncode == 0:
                                 log.debug("Successfully created " + str(message.Mailbag_Message_ID) + ".pdf")
                             else:
                                 if stdout:
-                                    log.warn("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stdout))
+                                    log.debug("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stdout.decode("utf-8"))
                                 if stderr:
-                                    desc = "Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stderr)
-                                    errors = helper.handle_error(errors, None, desc, "error")
+                                    desc = "Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + stderr.decode("utf-8")
+                                    errors = helper.handle_error(errors, None, desc, "warn")
                             # delete the HTML file
                             if os.path.isfile(pdf_name):
                                 os.remove(html_name)
@@ -106,7 +106,7 @@ if not skip_registry:
                             errors = helper.handle_error(errors, e, desc)
 
             except Exception as e:
-                desc = "Error creating PDF derivative with chrome"
+                desc = "Error creating PDF derivative"
                 errors = helper.handle_error(errors, e, desc)
 
             message.Error.extend(errors["msg"])
