+++ mailbagit/globals.py
